requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
filelock==3.0.12
flake8==3.8.4
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
h11==0.12.0
identify==2.1.3
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.5.0
orjson==3.5.1
packaging==20.9
passlib==1.7.4
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.11.1
promise==2.3
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.1
pyflakes==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.4.0b2
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.13.4
virtualenv==20.4.2
watchgod==0.7
websockets==8.1
